requirements.txt
flask
requests
pandas
openpyxl
urllib3
Office365-REST-Python-Client